requirements.txt
Flask==3.0.3
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
uvloop==0.20.0 ; platform_system != "Windows"